Create a Dockerfile and update Makefile and README.
Since it needs the NVML runtime it's based on CUDA image and needs
nvidia-docker. The whole CUDA is not needed, so we could possibly optimize it.

It needs the compiler to build the psutil Python package.
We could use some build image and not put the compiler in the target image.

We mount the master's config file from host into the container.