Fix error when invoking on MIG enabled gpus
`nvidia-smi -L` output for GPUs that are MIG enabled looks different
which results in some output lines in `gpu_info` to not match. This got
fixed by filtering out None matches.

Additionally `nvidia-smi` and `pynvml` can not get the utilization of a
GPU for which MIG devices are enabled. This results in a 'NVMLError' in
`device_status` which has to be catched. The utilization then is getting
set to 'N/A'.